chore: move validation error into utils package
