--- packages/lens/src/errors.ts
@@ -1,12 +1,0 @@
-export class ValidationNotValid extends Error {
-  constructor(input: string | Error) {
-    if (input instanceof Error) {
-      super(input.message)
-      Object.assign(this, input)
-
-      this.name = 'ValidationNotValid'
-    } else {
-      super(input)
-    }
-  }
-}
