+++ packages/utils/src/errors/index.ts
@@ -1,1 +1,4 @@
-export { PluginActionNotImplementedError } from './plugin'
+export {
+  PluginActionNotImplementedError,
+  ValidationNotValid,
+} from './plugin'
