+++ packages/utils/src/errors/plugin.ts
@@ -4,3 +4,16 @@ export class PluginActionNotImplementedError extends Error {
     this.name = 'PluginActionNotImplementedError'
   }
 }
+
+export class ValidationNotValid extends Error {
+  constructor(input: string | Error) {
+    if (input instanceof Error) {
+      super(input.message)
+      Object.assign(this, input)
+
+      this.name = 'ValidationNotValid'
+    } else {
+      super(input)
+    }
+  }
+}
